requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.11.0
pydantic-settings==2.8.1
pyjwt==2.10.0
pandas==2.2.3
pyarrow==18.1.0
requests==2.32.3
beautifulsoup4==4.12.3
shapely>=2.0
python-dotenv==1.0.1
